# backend/tests/snapshots/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Details contained in the top-level error envelope.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "description": "Stable API error envelope.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "NextPlanAction": {
        "enum": [
          "continue",
          "adjust"
        ],
        "title": "NextPlanAction",
        "type": "string"
      },
...
      "ReplanMode": {
        "enum": [
          "initial",
          "continue",
          "adjust"
        ],
        "title": "ReplanMode",
        "type": "string"
      },
...
      "ReviewCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "adjustment_request": {
            "anyOf": [
              {
                "maxLength": 300,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjustment Request"
          },
          "blockers": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockers"
          },
          "free_text": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text"
          },
          "mood": {
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Mood",
            "type": "integer"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "review_date": {
            "format": "date",
            "title": "Review Date",
            "type": "string"
          }
        },
        "required": [
          "plan_id",
          "review_date",
          "mood"
        ],
        "title": "ReviewCreateRequest",
        "type": "object"
      },
      "ReviewListResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "ReviewListResponse",
        "type": "object"
      },
      "ReviewResponse": {
        "additionalProperties": false,
        "properties": {
          "abandoned_count": {
            "minimum": 0.0,
            "title": "Abandoned Count",
            "type": "integer"
          },
          "adjustment_request": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adjustment Request"
          },
          "blockers": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockers"
          },
          "companion_message": {
            "title": "Companion Message",
            "type": "string"
          },
          "completed_count": {
            "minimum": 0.0,
            "title": "Completed Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "free_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Free Text"
          },
          "mood": {
            "title": "Mood",
            "type": "integer"
          },
          "next_plan_action": {
            "$ref": "#/components/schemas/NextPlanAction"
          },
          "next_plan_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Plan Run Id"
          },
          "plan_id": {
            "format": "uuid",
            "title": "Plan Id",
            "type": "string"
          },
          "replan_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Replan Reason"
          },
          "review_date": {
            "format": "date",
            "title": "Review Date",
            "type": "string"
          },
          "review_id": {
            "format": "uuid",
            "title": "Review Id",
            "type": "string"
          },
          "suggested_replan": {
            "title": "Suggested Replan",
            "type": "boolean"
          }
        },
        "required": [
          "review_id",
          "plan_id",
          "review_date",
          "mood",
          "blockers",
          "adjustment_request",
          "free_text",
          "completed_count",
          "abandoned_count",
          "suggested_replan",
          "replan_reason",
          "next_plan_action",
          "companion_message",
          "next_plan_run_id",
          "created_at"
        ],
        "title": "ReviewResponse",
        "type": "object"
      },
...
      "RunStatus": {
        "enum": [
          "pending",
          "running",
          "completed",
          "degraded",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "type": "string"
      },
...
      "StartNextPlanResponse": {
        "additionalProperties": false,
        "properties": {
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "replan_mode": {
            "$ref": "#/components/schemas/ReplanMode"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status",
          "replan_mode",
          "events_url"
        ],
        "title": "StartNextPlanResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Career Planning Buddy",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reviews": {
      "get": {
        "operationId": "list_reviews_api_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "plan_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Plan Id"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reviews",
        "tags": [
          "reviews"
        ]
      },
      "post": {
        "operationId": "create_review_api_v1_reviews_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Review",
        "tags": [
          "reviews"
        ]
      }
    },
    "/api/v1/reviews/{review_id}/start-next-plan": {
      "post": {
        "operationId": "start_next_plan_api_v1_reviews__review_id__start_next_plan_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartNextPlanResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Next Plan",
        "tags": [
          "reviews"
        ]
      }
    },
...
  }
}